+++ CHANGELOG.md
@@ -16,7 +16,8 @@
   * Non-delayed reaction when a pre-assigned tasks cannot be run because of time request
 * Resource weights (configured via `--weight`). They are used to adjust priority between different tasks or between
   multiple resource request variants within the same task.
-* Min-utilization is also utilized in main scheduler (not only autoalloc scheduler)
+* Worker can be started with `--min-utilization` (so not only through autoalloc).
+  Min-utilization is now respected also in main scheduler (no only autoalloc scheduler)
 
 
 ## v0.25.1